Include summary in help output for individual commands if we have it
Sometimes the "summary" is all the help that is required for a command and adding more in "Help" would redundant.
Also, sometimes "summary" serves as a useful introduction to a longer  "Help" document. 

So when people ask for the detailed "help" for a command, include the "Summary" as well as the full help.  

It's the same value as including a title when someone asks for a document.